server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["XDeviation <xth9681@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.12"
httpx = "^0.28.1"
pydantic = {extras = ["email"], version = "^2.11.3"}
gunicorn = "^23.0.0"
uvicorn = "^0.34.2"
pyjwt = "^2.10.1"
dotenv = "^0.9.9"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
pyyaml = "^6.0.2"
motor = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
